get webpack hot reload to work
html-loader extract-loader file-loader combo doesn't play well with
webpack devserver, unable to have a html file as entry point and load JS
file from html file. doing this as a work around so that it is possible
to have hot reload when the html is editted on the fly.

but it takes 4 secs for it to hot reload :(